Rename selector to use atom-* elements
Atom changed the way they layout their editor dom which broke the theme a bit.
Text colors (white on white) were the main things that I saw. These fixes
address some of those bugs. I can't find one single PR that references this, but
it seems to be related to:

* atom/atom#3633
* atom/atom#4248
* atom/atom#4955